+++ Cesium3DTilesContent/include/Cesium3DTilesContent/SubtreeAvailability.h
@@ -419,6 +419,7 @@ private:
       uint32_t relativeTileLevel,
       uint64_t relativeTileMortonId,
       AvailabilityView& availabilityView,
+      Cesium3DTiles::Availability& availability,
       bool isAvailable) noexcept;
 
   bool isAvailableUsingBufferView(