+++ Cesium3DTilesContent/test/TestSubtreeAvailability.cpp
@@ -1023,6 +1023,7 @@ TEST_CASE("SubtreeAvailability modifications") {
   REQUIRE(maybeAvailability);
 
   SubtreeAvailability& availability = *maybeAvailability;
+  const Subtree& subtree = availability.getSubtree();
 
   SUBCASE("initially has all tiles available, and no content or subtrees "
           "available") {
@@ -1063,6 +1064,9 @@ TEST_CASE("SubtreeAvailability modifications") {
         QuadtreeTileID(0, 0, 0),
         QuadtreeTileID(4, 15, 15)));
 
+    CHECK_UNARY(subtree.tileAvailability.bitstream);
+    CHECK_UNARY_FALSE(subtree.tileAvailability.constant);
+
     availability.setContentAvailable(
         QuadtreeTileID(0, 0, 0),
         QuadtreeTileID(4, 15, 15),
@@ -1078,6 +1082,9 @@ TEST_CASE("SubtreeAvailability modifications") {
         QuadtreeTileID(4, 15, 15),
         0));
 
+    CHECK_UNARY(subtree.contentAvailability[0].bitstream);
+    CHECK_UNARY_FALSE(subtree.contentAvailability[0].constant);
+
     availability.setSubtreeAvailable(
         QuadtreeTileID(0, 0, 0),
         QuadtreeTileID(5, 31, 31),
@@ -1089,5 +1096,8 @@ TEST_CASE("SubtreeAvailability modifications") {
     CHECK(availability.isSubtreeAvailable(
         QuadtreeTileID(0, 0, 0),
         QuadtreeTileID(5, 31, 31)));
+
+    CHECK_UNARY(subtree.childSubtreeAvailability.bitstream);
+    CHECK_UNARY_FALSE(subtree.childSubtreeAvailability.constant);
   }
 }